requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via starlette
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.29.0
    # via -r requirements.in
bcrypt==4.0.1
    # via
    #   -r requirements.in
    #   passlib
build==1.0.3
    # via poetry
cachecontrol[filecache]==0.13.1
    # via
    #   cachecontrol
    #   poetry
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via xattr
charset-normalizer==3.3.2
    # via requests
cleo==2.1.0
    # via poetry
click==8.1.7
    # via uvicorn
crashtest==0.4.1
    # via
    #   cleo
    #   poetry
distlib==0.3.8
    # via virtualenv
dnspython==2.5.0
    # via email-validator
dulwich==0.21.7
    # via poetry
ecdsa==0.18.0
    # via python-jose
email-validator==2.1.0.post1
    # via pydantic
fastapi==0.109.0
    # via -r requirements.in
fastjsonschema==2.19.1
    # via poetry
filelock==3.13.1
    # via
    #   cachecontrol
    #   virtualenv
greenlet==3.0.3
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   email-validator
    #   requests
importlib-metadata==7.0.1
    # via keyring
installer==0.7.0
    # via poetry
jaraco-classes==3.3.0
    # via keyring
keyring==24.3.0
    # via poetry
mako==1.3.2
    # via alembic
markupsafe==2.1.4
    # via mako
more-itertools==10.2.0
    # via jaraco-classes
msgpack==1.0.7
    # via cachecontrol
packaging==23.2
    # via
    #   build
    #   poetry
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pexpect==4.9.0
    # via poetry
pkginfo==1.9.6
    # via poetry
platformdirs==3.11.0
    # via
    #   poetry
    #   virtualenv
poetry==1.7.1
    # via
    #   -r requirements.in
    #   poetry-plugin-export
poetry-core==1.8.1
    # via
    #   poetry
    #   poetry-plugin-export
poetry-plugin-export==1.6.0
    # via poetry
ptyprocess==0.7.0
    # via pexpect
pyasn1==0.5.1
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic[email]==2.6.0
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.16.1
    # via pydantic
pydantic-settings==2.1.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via
    #   build
    #   poetry
python-dotenv==1.0.1
    # via pydantic-settings
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
rapidfuzz==3.6.1
    # via cleo
requests==2.31.0
    # via
    #   cachecontrol
    #   poetry
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via poetry
rsa==4.9
    # via python-jose
shellingham==1.5.4
    # via poetry
six==1.16.0
    # via ecdsa
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.25
    # via
    #   -r requirements.in
    #   alembic
starlette==0.35.1
    # via fastapi
tomlkit==0.12.3
    # via poetry
trove-classifiers==2024.1.31
    # via poetry
typing-extensions==4.9.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.2.0
    # via
    #   dulwich
    #   requests
uvicorn==0.27.0.post1
    # via -r requirements.in
virtualenv==20.25.0
    # via poetry
xattr==0.10.1
    # via poetry
zipp==3.17.0
    # via importlib-metadata
